server/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==1.2.2
google-genai==1.66.0
Pillow==12.1.1
python-multipart==0.0.22
pydantic==2.12.5
pygments==2.19.2

# Database (Async)
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0